fix incorrect function definitions
In 8e63251 (add parent button to the editor), the version of
add_save_button in static/js/helpers.js was renamed to
add_save_buttons, while the version in server/assets/helpers.js and
call-site was updated to call add_server_buttons.

At the same time, the version of set_save_button in
static/js/helpers.js was renamed to set_server_button, but no
call-site was changed. This smells like fat fingers ;)