server: add read only mode
In read only mode the endpoint to save new effects or add new versions
is deactivated. Also in the gallery the create button is deactivated and
there is a notice that says:

    The server is in maintenance mode. You can not create or modify effects.

The button to save effects in the editor is not deactivated as it will
take a bit more time and wanted to have the page up as soon as possible.